requirements.txt
django==2.*
djangorestframework==3.8.2
mysqlclient==1.3.14